fix for https://github.com/nzjess/statsd-over-slf4j/issues/4.  changed API to take long for millis values

+++ src/main/java/org/ubercraft/statsd/StatsdLogger.java
@@ -4,13 +4,13 @@ public interface StatsdLogger extends StatsdCounter, StatsdTimer {
 
     // Info level
 
-    void infoStat(StatsdStatType type, int value, double sampleRate);
+    void infoStat(StatsdStatType type, long value, double sampleRate);
 
     // Debug level
 
-    void debugStat(StatsdStatType type, int value, double sampleRate);
+    void debugStat(StatsdStatType type, long value, double sampleRate);
 
     // Trace level
 
-    void traceStat(StatsdStatType type, int value, double sampleRate);
+    void traceStat(StatsdStatType type, long value, double sampleRate);
 }
